Minor refactoring of entrypoint to sfc, naming standardization
