+++ client/src/components/NewUserWelcome/NewUserWelcome.vue
