--- client/src/components/NewUserWelcome/index.js
@@ -1,6 +1,0 @@
-import NewUserWelcome from "./newUserWelcome";
-// import NewUserDict from "./welcomePageConfig.json";
-export default {
-    NewUserWelcome,
-    // NewUserDict
-};
