+++ client/src/entry/analysis/AnalysisRouter.js
@@ -53,7 +53,7 @@ import Confirmation from "components/login/Confirmation.vue";
 import DatasetDetails from "components/Details/DatasetDetails.vue";
 import Libraries from "components/Libraries";
 import { mountVueComponent } from "utils/mountVueComponent";
-import NewUserWelcome from "components/NewUserWelcome/newUserWelcome";
+import NewUserWelcome from "components/NewUserWelcome/NewUserWelcome.vue";
 
 import { newUserDict } from "../../../../static/plugins/welcome_page/new_user/static/topics/index";
